testing release area and remove cyclic imports
